refactorando: renomeando nome dos arquivos de pagina para atender o padrão
